refactor resizeElements out from App
